Add a remote authentication handler base
