+++ src/Microsoft.AspNet.Authentication/project.json
@@ -18,7 +18,15 @@
         "Microsoft.Extensions.WebEncoders": "1.0.0-*"
     },
     "frameworks": {
-        "dnx451": { },
-        "dnxcore50": { }
+        "dnx451": {
+            "frameworkAssemblies": {
+                "System.Net.Http": ""
+            }
+        },
+        "dnxcore50": {
+            "dependencies": {
+                "System.Net.Http": "4.0.1-beta-*"
+            }
+        }
     }
 }
